Move helper functions to user
* Move two helper functions with a single user to its module.
* Remove unused `branch_of_daily_build`
